添加了一些东西。

+++ Calculator_OC/Calculator.h
@@ -25,7 +25,7 @@
  *
  *  @return 返回结算结果
  */
-- (long double)computeExpression:(NSString *)expression;
+- (double)computeExpression:(NSString *)expression;
 
 
 @end
